requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pymupdf4llm==0.0.12
pydantic-ai==0.0.14
smolagents==1.1.0
google-generativeai==0.8.3
faiss-cpu==1.8.0
sentence-transformers==3.0.1
python-multipart==0.0.9
structlog==24.1.0
pytest==8.0.0
python-dotenv==1.0.1